Moved helper file and added docstrings fallback
